# 백엔드/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HUFS Town API
  version: 0.5.0
  description: GDG HUFS SSO, spaces, invitations, map editing and world admission. Media signaling uses the authenticated world WebSocket contract; Node control routes are private. Never log credentials, invite codes or admission headers.
paths:
...
  /api/v1/spaces/{id}:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    get:
      operationId: getSpace
      security: [{ sessionCookie: [] }]
      description: PRIVATE requires membership; a denied read is indistinguishable from a missing space (404). UNLISTED can be read by ID by any active signed-in account.
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
    patch:
      operationId: updateSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Existing members retain access after visibility changes. Capacity and map template are immutable in this version.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpaceEdit" }
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
...
  /api/v1/spaces/{spaceId}/scheduled-events:
    parameters: [{ $ref: "#/components/parameters/SpaceAssetSpaceId" }]
    get:
      operationId: listScheduledSpaceEvents
      security: [{ sessionCookie: [] }]
      description: Lists upcoming events and cancellations from the last 30 days. PUBLIC and UNLISTED spaces are readable unless the account is access-blocked; PRIVATE schedules require membership.
      responses:
        "200":
          description: At most 50 upcoming event schedules.
          content:
            application/json:
              schema:
                {
                  type: array,
                  items: { $ref: "#/components/schemas/ScheduledEvent" },
                }
        default: { $ref: "#/components/responses/SpaceError" }
    post:
      operationId: createScheduledSpaceEvent
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER and ADMIN only. Starts 1 minute to 365 days ahead; optional end is 1 minute to 12 hours after start. HTTPS materials only, max 50 upcoming schedules per space and 10 created per account per day.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduledEventDraft" }
      responses:
        "200":
          description: Created schedule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduledEvent" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{spaceId}/scheduled-events/{eventId}:
    parameters:
      - { $ref: "#/components/parameters/SpaceAssetSpaceId" }
      - name: eventId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateScheduledSpaceEvent
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER and ADMIN only. Only future, non-cancelled schedules can be updated.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduledEventDraft" }
      responses:
        "200":
          description: Updated schedule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduledEvent" }
        default: { $ref: "#/components/responses/SpaceError" }
    delete:
      operationId: cancelScheduledSpaceEvent
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER and ADMIN only. Cancellation is retained so attendees can see the notice.
      responses:
        "200":
          description: Cancelled schedule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduledEvent" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{spaceId}/scheduled-events/{eventId}/rsvp:
    parameters:
      - { $ref: "#/components/parameters/SpaceAssetSpaceId" }
      - name: eventId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: respondToScheduledEvent
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Any active account that can read the space may respond GOING, INTERESTED, or DECLINED to a future event. Only aggregate counts and the caller's own response are returned; this expresses intent and does not reserve a seat or grant space admission.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduledEventRsvpDraft" }
      responses:
        "200":
          description: Updated schedule and aggregate RSVP counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduledEvent" }
        default: { $ref: "#/components/responses/SpaceError" }
    delete:
      operationId: clearScheduledEventResponse
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Removes the caller's response. This is allowed after cancellation or start so the user can withdraw their saved preference.
      responses:
        "200":
          description: Updated schedule and aggregate RSVP counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduledEvent" }
        default: { $ref: "#/components/responses/SpaceError" }
...
components:
...
  parameters:
    SpaceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    SpaceAssetSpaceId:
      name: spaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: HUFS_TOWN_SESSION }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-TOKEN }
  schemas:
...
    ScheduledEventDraft:
      type: object
      required:
        [title, description, instructions, resourceUrl, startsAt, endsAt]
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 280 }
        instructions: { type: string, maxLength: 500 }
        resourceUrl:
          { type: string, maxLength: 512, description: Empty or HTTPS URL. }
        startsAt: { type: string, format: date-time }
        endsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: string, maxLength: 0 }
    ScheduledEvent:
      type: object
      required:
        [
          id,
          title,
          description,
          instructions,
          resourceUrl,
          startsAt,
          endsAt,
          cancelled,
          createdAt,
          updatedAt,
          goingCount,
          interestedCount,
          declinedCount,
          myResponse,
        ]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 80 }
        description: { type: string, maxLength: 280 }
        instructions: { type: string, maxLength: 500 }
        resourceUrl: { type: string, maxLength: 512 }
        startsAt: { type: string, format: date-time }
        endsAt:
          description: Empty when no end time is set.
          oneOf:
            - { type: string, format: date-time }
            - { type: string, maxLength: 0 }
        cancelled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        goingCount: { type: integer, minimum: 0 }
        interestedCount: { type: integer, minimum: 0 }
        declinedCount: { type: integer, minimum: 0 }
        myResponse: { type: string, enum: ["", GOING, INTERESTED, DECLINED] }
    ScheduledEventRsvpDraft:
      type: object
      required: [response]
      additionalProperties: false
      properties:
        response: { type: string, enum: [GOING, INTERESTED, DECLINED] }
...
    SpaceEdit:
      type: object
      required: [name, description, visibility]
      properties:
        name:
          {
            type: string,
            minLength: 1,
            maxLength: 60,
            description: Trimmed; no control/format characters.,
          }
        description:
          {
            type: string,
            maxLength: 300,
            description: Trimmed single-line description; no control/format characters.,
          }
        visibility: { type: string, enum: [PUBLIC, UNLISTED, PRIVATE] }
        approvalRequired:
          {
            type: boolean,
            default: false,
            description: Only applies to PUBLIC or UNLISTED spaces. Existing members and valid invitees can enter without owner approval.,
          }
...
    Space:
      allOf:
        - { $ref: "#/components/schemas/SpaceEdit" }
        - type: object
          required:
            [
              id,
              capacity,
              templateId,
              role,
              approvalRequired,
              joinRequestStatus,
              favorite,
              archived,
            ]
          properties:
            id: { type: string, format: uuid }
            capacity: { type: integer, minimum: 1, maximum: 100 }
            templateId:
              {
                type: string,
                enum: [CAMPUS_SQUARE, STUDY_SPACE, MEETUP_HALL, OFFICE],
              }
            role: { type: string, enum: [OWNER, ADMIN, MEMBER, ""] }
            approvalRequired: { type: boolean }
            joinRequestStatus:
              { type: string, enum: [PENDING, APPROVED, REJECTED, ""] }
            favorite:
              {
                type: boolean,
                description: Whether the authenticated user saved this space.,
              }
            archived:
              {
                type: boolean,
                description: True when returned by the owner's archived-space view.,
              }
...
    AuthError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
  responses:
...
    Space:
      description: Space metadata and the caller's role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Space" }
...
    SpaceError:
      description: 400 invalid input/invite/event date; 401 session missing; 403 owner/account/CSRF/approval/access-block denied; 404 missing or private space; 409 quota/request/event/ownership-transfer/archive conflict; 429 per-account rate or daily event limit; 503 storage unavailable. Uses the common code/message shape.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }
...